fix(vo): 修改收支统计年份视图对象类型ID数据类型

+++ src/main/java/top/aiolife/record/pojo/vo/IncStaByYearVO.java
@@ -13,7 +13,7 @@ public class IncStaByYearVO {
 
     private BigDecimal amt;
 
-    private Integer typeId;
+    private Long typeId;
 
     private String typeName;
 }
